+++ lib/karafka/schemas/consumer_group_topic.rb
@@ -12,7 +12,6 @@ module Karafka
       required(:max_bytes_per_partition).filled(:int?, gteq?: 0)
       required(:start_from_beginning).filled(:bool?)
       required(:batch_consuming).filled(:bool?)
-      required(:persistent).filled(:bool?)
     end
   end
 end