+++ spec/lib/karafka/connection/message_delegator_spec.rb
@@ -23,7 +23,6 @@ RSpec.describe Karafka::Connection::MessageDelegator do
     Karafka::Routing::Builder.instance.draw do
       topic :topic_name1 do
         consumer Class.new(Karafka::BaseConsumer)
-        persistent false
       end
     end